코드 정리 및 Directory, DirectoryInfo 구현 중

+++ Source/TGON/String/Encoding.cpp
@@ -158,7 +158,7 @@ int32_t Encoding::Convert(const Encoding& srcEncoding, const Encoding& destEncod
 
 int32_t Encoding::Convert(const Encoding& srcEncoding, const Encoding& destEncoding, const gsl::span<std::byte>& srcBytes, const gsl::span<std::byte>& destBytes)
 {
-    return Convert(srcEncoding, destEncoding, &srcBytes[0], srcBytes.size(), &destBytes[0], destBytes.size());
+    return Convert(srcEncoding, destEncoding, &srcBytes[0], static_cast<int32_t>(srcBytes.size()), &destBytes[0], static_cast<int32_t>(destBytes.size()));
 }
 
 std::vector<char32_t> Encoding::GetChars(const std::byte* bytes, int32_t count) const
@@ -189,7 +189,7 @@ std::vector<char32_t> Encoding::GetChars(const std::byte* bytes, int32_t count) 
 
 std::vector<char32_t> Encoding::GetChars(const gsl::span<std::byte>& bytes) const
 {
-    return GetChars(&bytes[0], bytes.size());
+    return GetChars(&bytes[0], static_cast<int32_t>(bytes.size()));
 }
 
 int32_t Encoding::GetCharCount(const std::byte* bytes, int32_t count) const
@@ -220,7 +220,7 @@ int32_t Encoding::GetCharCount(const std::byte* bytes, int32_t count) const
 
 int32_t Encoding::GetCharCount(const gsl::span<std::byte>& bytes) const
 {
-    return GetCharCount(&bytes[0], bytes.size());
+    return GetCharCount(&bytes[0], static_cast<int32_t>(bytes.size()));
 }
 
 const std::string_view& Encoding::GetEncodingName() const noexcept
